move HerderSCPDriver to separate file
